chore: Restruct folders. Restruct folder in modules following DDD concepts.
