+++ .eslintrc.json
@@ -30,7 +30,14 @@
           "format" : null
        }
      ],
-      "@typescript-eslint/no-unused-vars": ["error",{
+     "@typescript-eslint/naming-convention": [
+      "error",
+      {
+        "selector": "interface",
+        "prefix": ["I"],
+        "format": ["PascalCase"]
+      }
+    ], "@typescript-eslint/no-unused-vars": ["error",{
         "argsIgnorePattern":"_"
       }],
       "import/extensions":[
