+++ src/modules/users/infra/http/middlewares/ensureAuthenticated.ts
@@ -1,8 +1,8 @@
 import { Request, Response, NextFunction } from 'express';
 import { verify } from 'jsonwebtoken';
 
-import authConfig from '../config/auth';
-import AppError from '../error/AppError';
+import authConfig from '@config/auth';
+import AppError from '@shared/error/AppError';
 
 interface TokenPayload {
   iat: number;
