+++ src/modules/users/infra/http/routes/sessions.routes.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 import 'express-async-errors';
 
-import AuthenticateUserService from '../services/AuthenticateUserService';
+import AuthenticateUserService from '@modules/users/services/AuthenticateUserService';
 
 const sessionsRouter = Router();
