+++ src/modules/users/infra/http/routes/users.routes.ts
@@ -3,10 +3,10 @@ import 'express-async-errors';
 
 import multer from 'multer';
 
-import uploadConfig from '../config/upload';
-import CreateUserService from '../services/CreateUserService';
-import ensureAuthenticated from '../middlewares/ensureAuthenticated';
-import UptadeUserAvatarService from '../services/UpdateUserAvatarService';
+import uploadConfig from '@config/upload';
+import CreateUserService from '@modules/users/services/CreateUserService';
+import ensureAuthenticated from '@modules/users/infra/http/middlewares/ensureAuthenticated';
+import UptadeUserAvatarService from '@modules/users/services/UpdateUserAvatarService';
 
 const usersRouter = Router();
