+++ src/modules/users/services/AuthenticateUserService.ts
@@ -2,9 +2,9 @@ import { getRepository } from 'typeorm';
 import { compare } from 'bcryptjs';
 import { sign } from 'jsonwebtoken';
 
-import User from '../models/User';
-import authConfig from '../config/auth';
-import AppError from '../error/AppError';
+import authConfig from '@config/auth';
+import AppError from '@shared/error/AppError';
+import User from '../infra/typeorm/entities/User';
 
 interface Request {
   email: string;
