+++ src/modules/users/services/UpdateUserAvatarService.ts
@@ -3,9 +3,9 @@ import { getRepository } from 'typeorm';
 import path from 'path';
 import fs from 'fs';
 
-import uploadConfig from '../config/upload';
-import User from '../models/User';
-import AppError from '../error/AppError';
+import uploadConfig from '@config/upload';
+import User from '@modules/users/infra/typeorm/entities/User';
+import AppError from '@shared/error/AppError';
 
 interface Request {
   user_id: string;
